refactor common into MergeUtils
