+++ src/main/scala/com/ebiznext/comet/utils/conversion/BigQueryUtils.scala
@@ -21,4 +21,30 @@ object BigQueryUtils {
   def bqSchema(schema: DataType): BQSchema = {
     BigQuerySchemaConverters.toBigQuerySchema(schema.asInstanceOf[StructType])
   }
+
+  /** Spark BigQuery driver consider integer in BQ as Long. We need to convert the Int DataType to
+    * LongType before loading the data. As a good practice, always use long when dealing with big
+    * query in your YAML Schema.
+    * @param schema
+    * @return
+    */
+  def normalizeSchema(schema: StructType): StructType = {
+    val fields = schema.fields.map { field =>
+      field.dataType match {
+        case dataType: StructType =>
+          field.copy(dataType = normalizeSchema(dataType))
+        case ArrayType(elementType: StructType, nullable) =>
+          field.copy(dataType = ArrayType(normalizeSchema(elementType), nullable))
+        case ArrayType(_: IntegerType, nullable) =>
+          field.copy(dataType = ArrayType(LongType, nullable))
+        case IntegerType => field.copy(dataType = LongType)
+        case _           => field
+      }
+    }
+    schema.copy(fields = fields)
+  }
+  ///////////////////////////////////////////////////////////////////////////
+  // Merge From BigQuery Data Source
+  ///////////////////////////////////////////////////////////////////////////
+
 }
